chore: move navbar/ folder to features/
