+++ src/app/page.tsx
@@ -1,14 +1,14 @@
 import About from "@/components/about/About";
 import ContactMe from "@/components/contact_me/ContactMe";
-import Header from "@/components/header/Header";
 import Hero from "@/components/hero/Hero";
-import Works from "@/components/projects/Works";
 import Skills from "@/components/skills/Skills";
+import Works from "@/components/works/Works";
+import Navbar from "@/features/navbar/Navbar";
 
 export default function Home() {
 	return (
 		<div className={"flex flex-col items-center text-[#fff] bg-[#0B1215]"}>
-			<Header />
+			<Navbar />
 			<Hero />
 			<About />
 			<Skills />
